==> 03_diversity_and_ordination.Rmd <==
---
title: "03_ordination_diversity"
author: "Daniel"
date: "27/09/2022"
output: github_document
#editor_options: 
  #chunk_output_type: console
---


```{r message=FALSE, warning=FALSE, include=FALSE}
rm(list=ls()) #I always start scripts this way to make sure I have a clean R environment
library('tidyr')
library('dplyr')
library('metacoder')
library('ggplot2')
library('agricolae')
library('vegan')
library('MicrobiotaProcess')
library('phyloseq')
library('ggtree')
library('coin')
library('Hmisc')

```
Note now that it is 'taxmap_object_binary.rds' - we have transformed all data to a binary presence/absence matrix.
Metabarcoding data although gives us the number of reads for each BIN, does not mean we have a measure of abundance for that BIN. Detectability issues and biases must be well understood. Refer to this article for better understanding: https://onlinelibrary.wiley.com/doi/full/10.1111/1755-0998.13703

```{r message = FALSE, warning = FALSE}

obj <- readRDS('data/correct_anal/taxmap_object_binary.rds') #loads the taxmap object created in script 02_metacoder_heat_trees

obj %>%  metacoder::filter_taxa(taxon_names %in% c("Lepidoptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> leps #we will create separate files for each order as this simplifies downstream analysis in microbiotaprocess - until I find a way to filter taxa on the mpse object directly.

obj %>%  metacoder::filter_taxa(taxon_names %in% c("Coleoptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> coleo
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Diptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> dips
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Hymenoptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> bees #i know its hymenoptera
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Hemiptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> hemi
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Blattodea"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> blats 

sample <- read.csv('data/final_anal/location_ctrl.csv')
```

The chunk above now gave us a filtered taxmap object for each of the Arthropod Program's focal Orders (plus Diptera because I think its interesting).
Downstream analysis will make use of these filtered data sets, but we continue with general info (whole data set) between seasons.

```{r simpleBoxplotANOVAbetweenSeason}
sample$inv_simp <- diversity(obj$data$tax_data[, sample$sampleID],
                             index = "invsimpson",
                             MARGIN = 2) # What orietation the matrix is in
ggplot(sample, aes(x = SEASON, y = inv_simp)) +
  geom_boxplot()

anova_result <- aov(inv_simp ~ SEASON, sample)
summary(anova_result)
tukey_result <- HSD.test(anova_result, "SEASON", group = TRUE)
print(tukey_result)

group_data <- tukey_result$groups[order(rownames(tukey_result$groups)),]

pdf("./03_diversity_and_ordination_files/boxplot.pdf")
ggplot(sample, aes(x = SEASON, y = inv_simp)) +
    geom_violin() + 
    geom_text(data = data.frame(),
            aes(x = rownames(group_data), y = max(sample$inv_simp) + 1, label = group_data$groups),
            col = 'black',
            size = 10) +
  stat_summary(fun.data = mean_sdl, geom = "errorbar", width = 0.1) +
  stat_summary(fun.data = mean_sdl, geom = "crossbar", width = 0.2, fatten = 2) +
  scale_fill_manual(values = c("wet" = "darkgoldenrod", "dry"="darkblue"))+
  ggtitle("Inverse Simpson diversity") +
  xlab("Season") +
  ylab("Inverse simpson index")
dev.off()
```

We can also see differences between sampling days or location (site) although there are none. 

```{r simpleBoxplotANOVAbetweenDays}
ggplot(sample, aes(x = day, y = inv_simp)) +
  geom_boxplot()
anova_result <- aov(inv_simp ~ day, sample)
summary(anova_result)
tukey_result_day <- HSD.test(anova_result, "day", group = TRUE)
print(tukey_result_day)

ggplot(sample, aes(x = site, y = inv_simp)) +
  geom_boxplot()
anova_result <- aov(inv_simp ~ site, sample)
summary(anova_result)
tukey_result_site <- HSD.test(anova_result, "site", group = TRUE)
print(tukey_result_site)

group_data_day <- tukey_result_day$groups[order(rownames(tukey_result_day$groups)),]
ggplot(sample, aes(x = day, y = inv_simp)) +
  geom_text(data = data.frame(),
            aes(x = rownames(group_data_day), y = max(sample$inv_simp) + 1, label = group_data_day$groups),
            col = 'black',
            size = 10) +
  geom_boxplot() +
  ggtitle("Inverse Simpson diversity") +
  xlab("Day") +
  ylab("Inverse simpson index")

group_data_site <- tukey_result_site$groups[order(rownames(tukey_result_site$groups)),]
ggplot(sample, aes(x = site, y = inv_simp)) +
  geom_text(data = data.frame(),
            aes(x = rownames(group_data_site), y = max(sample$inv_simp) + 1, label = group_data_site$groups),
            col = 'black',
            size = 10) +
  geom_boxplot() +
  ggtitle("Inverse Simpson diversity") +
  xlab("site") +
  ylab("Inverse simpson index")

```

We can also calculate several diversity metrics for sampling season and day using phyloseq but we must convert the taxmap object 'obj' to phyloseq 

```{r multipleDiversityIndices}
ps_obj <- metacoder::as_phyloseq(obj,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID")

#normally, I load every package I need at the start of the script, but this causes function masking problems (e.g. both MicrobiotaProcess and Metacoder have a 'as_phyloseq' function and you have to specify to R which package you want to use -  see above: metacoder::as_phyloseq - this means I am telling R to use the 'as_phyloseq' function from the metacoder package)

plot_richness(ps_obj, color = "SEASON", x = "site") #phyloseq function
plot_richness(ps_obj, color = "day", x = "SEASON")
```

Figures above OK but we can do better.
I am using the MicrobiotaProcess next

```{r MBPPlotsDiversityIndices}
alphaobj <- get_alphaindex(ps_obj)
p_alpha <- ggbox(alphaobj, geom="violin", factorNames="SEASON", indexNames = c('Observe', 'Shannon'),  signifmap = TRUE) +
  scale_fill_manual(values=c("goldenrod", "steelblue"))+
  theme(strip.background = element_rect(colour=NA, fill="grey"))
pdf("./03_diversity_and_ordination_files/diversity_indices_season_binary.pdf")
p_alpha
dev.off()
```
And look at rarefaction curves for both seasons:
These are rarefaction curves of species (Y) and number of reads (X) - tells you how many read coverage you need to capture most diversity.

We can also do a Principal Coordinate Analysis to see not only the differences between seasons, but what BINs are the main drivers of these differences

```{r MBPPCoAPlots}
# distmethod
# "unifrac",  "wunifrac", "manhattan", "euclidean", "canberra", "bray", "kulczynski" ...(vegdist, dist)
pcoares <- get_pcoa(obj=ps_obj, distmethod="bray", method="hellinger")
# Visualizing the result
pcoaplot1 <- ggordpoint(obj=pcoares, biplot=TRUE, speciesannot=TRUE,
                        factorNames=c("SEASON"), ellipse=TRUE) +
  scale_color_manual(values=c("goldenrod", "steelblue")) +
  scale_fill_manual(values=c("goldenrod", "steelblue"))
# first and third principal co-ordinates
pcoaplot2 <- ggordpoint(obj=pcoares, pc=c(1, 3), biplot=TRUE, speciesannot=TRUE,
                        factorNames=c("SEASON"), ellipse=TRUE) +
  scale_color_manual(values=c("goldenrod", "steelblue")) +
  scale_fill_manual(values=c("goldenrod", "steelblue"))

pdf("./03_diversity_and_ordination_files/pcoa_2_axis_binary.pdf")
pcoaplot1 
dev.off()

pdf("./03_diversity_and_ordination_files/pcoa_3_axis.pdf")
pcoaplot2
dev.off()

pcoaplot1 | pcoaplot2
```

Next we will see which are the most frequent taxa (in terms of reads)
We can edit this polot with two important compontents: taxlevel and topn
taxlevel means the level of taxonomic resolution for the filtering kingdom:phyllum:class:order:family:genus:species
topn mean the number of the most abundant taxa you want to include e.g. top 10 - be aware that the color may mess up if you choose a large number

```{r MBPTaxonAbundanceTOP10}
classtaxa <- get_taxadf(obj=ps_obj, taxlevel=4)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass <- ggbartax(obj=classtaxa, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("Presence (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(21))) +
  guides(fill= guide_legend(keyw0idth = 0.5, keyheight = 0.5))

#note the flag "count=TRUE", this shows now total reads, rather than proportion %
pclass2 <- ggbartax(obj=classtaxa, count=TRUE, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("Presence (total BINs)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./03_diversity_and_ordination_files/relative_abundance_top5_order_binary.pdf")
pclass
dev.off()

pdf("./03_diversity_and_ordination_files/count_reads_top5_order_binary.pdf")
pclass2
dev.off()

pclass | pclass2

classtaxa_fam <- get_taxadf(obj=ps_obj, taxlevel=5)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_fam <- ggbartax(obj=classtaxa_fam, facetNames="SEASON", topn=10) +
  xlab(NULL) +
  ylab("Presence (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(11,"Set3"))(51))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./03_diversity_and_ordination_files/relative_agundance_top20_family_binary.pdf")
pclass_fam
dev.off()

pclass | pclass_fam | pclass2
```


These following chunks might be best to compare metabarcoding vs traditional data but its OK to visualize below.
Way too many species difference between both sampling strategies and so the plots are messy  - after Kruskal, Wilcox and LDA, 264 taxa discriminate between wet and dry seasons.

```{r warning=FALSE}
#'coin' package used for kruskal and wilcox test
# Since the effect size was calculated by randomly re-sampling, 
# the seed should be set for reproducibly results.
set.seed(1024)
deres <- diff_analysis(obj = ps_obj, classgroup = "SEASON",
                       mlfun = "lda",
                       filtermod = "pvalue",
                       firstcomfun = "kruskal_test",
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       subclmin = 3,
                       subclwilc = TRUE,
                       secondalpha = 0.01,
                       lda=3,
                       action = "add")
deres
```

this figure gives you a cladogram of all the BINs (based on taxonomy) and highlight which species/genera/orders are different between treatments, in this case seasons

We can also create boxplots for families - this one which includes everything is very messy.
In the next script we repeat these analyses but for FOCAL orders only.

```{r BoxplotFIlteredFamily}
ps_obj %>% as.MPSE() %>% mp_rrarefy() %>% mp_diff_analysis(.abundance=RareAbundance, .group=SEASON, action='get') %>% dplyr::filter(grepl("^f__", f)) %>% ggdiffbox(colorlist=c("steelblue", "goldenrod"), notch = FALSE) -> ggdiffbox_family

pdf("./03_diversity_and_ordination_files/ggdiffbox_family_binary.pdf")
ggdiffbox_family
dev.off()

ps_obj %>% as.MPSE() %>% mp_rrarefy() %>% mp_diff_analysis(.abundance=RareAbundance, .group=SEASON, action='get') %>%
dplyr::filter(grepl("^f__", f)) %>% ggeffectsize(colorlist= c('#daa520', '#4682b4'), notch = FALSE) -> ggeffectsize_family

pdf("./03_diversity_and_ordination_files/ggeffectsize_family_binary.pdf")
ggeffectsize_family
dev.off()

ps_obj %>% as.MPSE() %>% mp_rrarefy() %>% mp_diff_analysis(.abundance=RareAbundance, .group=SEASON, action='get') %>% dplyr::filter(grepl("^o__", f)) %>% ggdiffbox(colorlist=c("goldenrod", "steelblue"), notch = FALSE) -> ggdiffbox_order

pdf("./03_diversity_and_ordination_files/ggdiffbox_order_binary.pdf")
ggdiffbox_order
dev.off()

```


###### Please pay attention to the code below as it breaks down the code above, but to speed up the process so we don't have to convert form PS to MPSE to diff_anal. Instead it creates the object ps_obj_rareMPSE which I then used to filter by family and do aeither ggdifbox or ggeffectsize

ps_obj %>%
  as.MPSE() %>%
  mp_diff_analysis(.abundance = RareAbundance, .group = SEASON, action = 'get') -> ps_obj_rareMPSE

ps_obj_rareMPSE %>% 
  dplyr::filter(grepl("^f__", f)) %>% 
  ggdiffbox(colorlist = c ("steelblue", "goldenrod"), notch = FALSE)

ps_obj_rareMPSE %>% 
  dplyr::filter(grepl("^f__", f)) %>% 
  ggeffectsize(colorlist = c ("steelblue", "goldenrod"), notch = FALSE)
