==> 04_focal_orders_seasons.Rmd <==
---
title: "04_focal_order_seasons"
author: "Daniel"
date: "06/10/2022"
output: github_document
#editor_options: 
  #chunk_output_type: console
---

```{r message=FALSE, warning=FALSE, include=FALSE}
rm(list=ls()) #I always start scripts this way to make sure I have a clean R environment
library('tidyr')
library('dplyr')
library('metacoder')
library('ggplot2')
library('agricolae')
library('vegan')
library('MicrobiotaProcess')
library('phyloseq')
library('ggtree')
library('coin')
```

```{r warning = FALSE, message=FALSE}

obj <- readRDS('data/correct_anal/taxmap_object_binary.rds') #loads the taxmap object created in script 02_metacoder_heat_trees

obj %>%  metacoder::filter_taxa(taxon_names %in% c("Lepidoptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> leps #we will create separate files for each order as this simplifies downstream analysis in microbiotaprocess - until I find a way to filter taxa on the mpse object directly.

obj %>%  metacoder::filter_taxa(taxon_names %in% c("Coleoptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> coleo
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Diptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> dips
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Hymenoptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> bees #i know its hymenoptera
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Hemiptera"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> hemi
obj %>%  metacoder::filter_taxa(taxon_names %in% c("Blattodea"),#here is to fliter the figure by groups
              subtaxa = TRUE) -> blats 

sample <- read.csv('data/correct_anal/location_ctrl.csv')
```

The script above gave me the taxmap object separated by order, however, it removes all taxonomic ranks up until "Order", so the script below is firstly, changing the taxmap object (e.g. leps) into a phyloseq object (as_phyloseq) and then using the microViz package to add the taxonomic ranks. These are necessary for some of the microbiotaprocess analyses otherwise the package complains that there is not enough taxonomic ranks.

There must be certainly a cleaner way how to do this without so many steps, but I then had to change the classification rank names again ("k__" instead of Kingdom) as this is more friendly with MicroBiotaProcess. Within the metabarcoding package community, there seems to have been several issues with the ranks, some have root, some don't, some have subfamily, some don't, so take note of the nomenclature you are using. I am using Kingdom;Phylum;Order;Class;Family;genus;species  (1 - 7).


```{r CreatingOrderFiles, message=FALSE, warning=FALSE, include=FALSE}
ps_leps <- metacoder::as_phyloseq(leps,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID") %>% microViz::tax_mutate(rank_0 = "Animalia",rank_01 = "Arthropoda", rank_02 = "Insecta", .before = 1) -> ps_leps 
colnames(tax_table(ps_leps)) <- c("k__", "p__", "c__", "o__", "f__", "g__", "s__")
  

ps_coleo <- metacoder::as_phyloseq(coleo,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID")  %>% microViz::tax_mutate(rank_0 = "Animalia",rank_01 = "Arthropoda", rank_02 = "Insecta", .before = 1) -> ps_coleo

colnames(tax_table(ps_coleo)) <- c("k__", "p__", "c__", "o__", "f__", "g__", "s__")

ps_dips <- metacoder::as_phyloseq(dips,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID")  %>% microViz::tax_mutate(rank_0 = "Animalia",rank_01 = "Arthropoda", rank_02 = "Insecta", .before = 1) -> ps_dips
colnames(tax_table(ps_dips)) <- c("k__", "p__", "c__", "o__", "f__", "g__", "s__")

ps_bees <- metacoder::as_phyloseq(bees,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID")  %>% microViz::tax_mutate(rank_0 = "Animalia",rank_01 = "Arthropoda", rank_02 = "Insecta", .before = 1) -> ps_bees
colnames(tax_table(ps_bees)) <- c("k__", "p__", "c__", "o__", "f__", "g__", "s__")

ps_blats <- metacoder::as_phyloseq(blats,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID")  %>% microViz::tax_mutate(rank_0 = "Animalia",rank_01 = "Arthropoda", rank_02 = "Insecta", .before = 1) -> ps_blats
colnames(tax_table(ps_blats)) <- c("k__", "p__", "c__", "o__", "f__", "g__", "s__")

ps_hemi <- metacoder::as_phyloseq(hemi,
                      otu_table = "tax_data",
                      otu_id_col = "bin_uri",
                      sample_data = sample,
                      sample_id_col = "sampleID")  %>% microViz::tax_mutate(rank_0 = "Animalia",rank_01 = "Arthropoda", rank_02 = "Insecta", .before = 1) -> ps_hemi
colnames(tax_table(ps_hemi)) <- c("k__", "p__", "c__", "o__", "f__", "g__", "s__")

```

Ok we now have datasets separated for each order (ForestGEO focal plus diptera - if we want more, we can adapt the script above to include more focal groups, or even to refine them - e.g. family level)
But below we will create now some taxonomic trees showing differences between seasons (as in the previous script).

```{r Lepidoptera}
set.seed(1024)
deresleps <- diff_analysis(obj = ps_leps, classgroup = "SEASON",
                       mlfun = "lda",
                       filtermod = "pvalue",
                       firstcomfun = "kruskal_test",
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       subclmin = 3,
                       subclwilc = TRUE,
                       secondalpha = 0.01,
                       lda=3,
                       action = "add")
deresleps

diffclade_leps <- ggdiffclade(
  obj=deresleps, 
  alpha=0.5, 
  linewd=0.01,
  skpointsize=0.05, 
  layout="radial",
  cladetext = 0.4,
  taxlevel=5, #taxonomy level from 1 to 7 kingdom:phylum:class:order:family:genus:species
  removeUnkown=TRUE,
  reduce=TRUE # This argument is to remove the branch of unknown taxonomy.
) +
  scale_fill_manual(
    values=c("goldenrod", "steelblue")
  ) +
  guides(color = guide_legend(
    keywidth = 0.1, 
    keyheight = 0.6,
    order = 5,
    ncol=3)
  ) +
  theme(
    panel.background=element_rect(fill=NA),
    legend.position="right", 
    plot.margin=margin(0,0,0,0),
    legend.spacing.y=unit(0.02, "cm"), 
    legend.title=element_text(size=7),
    legend.text=element_text(size=6), 
    legend.box.spacing=unit(0.02,"cm")
  )

pdf("./04_focal_orders_seasons_files/leps_cladogram.pdf")
diffclade_leps
dev.off()

classtaxa_leps <- get_taxadf(obj=ps_leps, taxlevel=5) #taxonomy level from 1 to 7 kingdom:phylum:class:order:family:genus:species
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_leps <- ggbartax(obj=classtaxa_leps, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("relative abundance (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./04_focal_orders_seasons_files/leps_relative_difference__families_season_binary.pdf")
pclass_leps
dev.off()

```

```{r Coleoptera}
set.seed(1024)
derescoleo <- diff_analysis(obj = ps_coleo, classgroup = "SEASON",
                       mlfun = "lda",
                       filtermod = "pvalue",
                       firstcomfun = "kruskal_test",
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       subclmin = 3,
                       subclwilc = TRUE,
                       secondalpha = 0.01,
                       lda=3,
                       action = "add")
derescoleo

classtaxa_coleo <- get_taxadf(obj=ps_coleo, taxlevel=5)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_coleo <- ggbartax(obj=classtaxa_coleo, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("relative abundance (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./04_focal_orders_seasons_files/coleo_diffclass_top5_binary.pdf")
pclass_coleo
dev.off()

es_coleo <- ggeffectsize(obj=derescoleo, 
                     lineheight=0.1,
                     linewidth=0.3, ytextsize = 0.3) + 
  scale_color_manual(values=c("goldenrod", "steelblue")) 

pdf("./04_focal_orders_seasons_files/coleo_effectsize.pdf")
es_coleo
dev.off()

```

```{r Diptera}
set.seed(1024)
deresdips <- diff_analysis(obj = ps_dips, classgroup = "SEASON",
                       mlfun = "lda",
                       filtermod = "pvalue",
                       firstcomfun = "kruskal_test",
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       subclmin = 3,
                       subclwilc = TRUE,
                       secondalpha = 0.01,
                       lda=3,
                       action = "add")
deresdips


classtaxa_dips <- get_taxadf(obj=ps_dips, taxlevel=5)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_dips <- ggbartax(obj=classtaxa_dips, facetNames="SEASON", topn=10) +
  xlab(NULL) +
  ylab("relative abundance (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./04_focal_orders_seasons_files/dips_top5_binary.pdf")
pclass_dips
dev.off()

es_dips <- ggeffectsize(obj=deresdips, 
                     lineheight=0.1,
                     linewidth=0.3, ytextsize = 0.3) + 
  scale_color_manual(values=c("goldenrod", "steelblue")) 

pdf("./04_focal_orders_seasons_files/diptera_effectsize.pdf")
es_dips
dev.off()

```


```{r Hemyptera}
set.seed(1024)
dereshemi <- diff_analysis(obj = ps_hemi, classgroup = "SEASON",
                       mlfun = "lda",
                       filtermod = "pvalue",
                       firstcomfun = "kruskal_test",
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       subclmin = 3,
                       subclwilc = TRUE,
                       secondalpha = 0.01,
                       lda=3,
                       action = "add")
dereshemi


classtaxa_hemi <- get_taxadf(obj=ps_hemi, taxlevel=5)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_hemi <- ggbartax(obj=classtaxa_hemi, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("relative abundance (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./04_focal_orders_seasons_files/hemiptera_top5_binary.pdf")
pclass_hemi
dev.off()

```

```{r Hymenoptera}

set.seed(1024)
deresbees <- diff_analysis(obj = ps_bees, classgroup = "SEASON",
                       mlfun = "lda",
                       filtermod = "pvalue",
                       firstcomfun = "kruskal_test",
                       firstalpha = 0.05,
                       strictmod = TRUE,
                       secondcomfun = "wilcox_test",
                       subclmin = 3,
                       subclwilc = TRUE,
                       secondalpha = 0.01,
                       lda=3,
                       action = "add")
deresbees


classtaxa_bees <- get_taxadf(obj=ps_bees, taxlevel=5)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_bees <- ggbartax(obj=classtaxa_bees, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("relative abundance (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./04_focal_orders_seasons_files/hymenoiptera_top5_bin.pdf")
pclass_bees
dev.off()
```

```{r blatts}
classtaxa_blats <- get_taxadf(obj=ps_blats, taxlevel=5)
# The 10 most abundant taxonomy will be visualized by default (parameter `topn=10`). 
pclass_blats <- ggbartax(obj=classtaxa_blats, facetNames="SEASON", topn=5) +
  xlab(NULL) +
  ylab("relative abundance (%)") +
  scale_fill_manual(values=c(colorRampPalette(RColorBrewer::brewer.pal(12,"Set3"))(31))) +
  guides(fill= guide_legend(keywidth = 0.5, keyheight = 0.5))

pdf("./04_focal_orders_seasons_files/blats_top5_binary.pdf")
pclass_blats
dev.off()


```
