==> 08_nmds_supplementary_material.Rmd <==
---
title: "08_nmds_supplementary_material"
author: "Daniel"
date: "14/02/2023"
output: github_document
editor_options:
  chunk_output_type: console
---
```{r}
rm(list=ls())
library("vegan")
library("mvabund")
library("dplyr")
library("elevatr")
library("ggplot2")
library("geosphere")
library("ape")
library("ggtree")

arthro <- read.csv('data/correct_anal/CYBs_correct_names.csv')
metadata <- read.csv('data/correct_anal/location_ctrl.csv')

arthro_trans <- t(arthro[, -2]) #transpose dataset (samples in rows, species in columns)
colnames(arthro_trans) <- arthro_trans[1,] #first row is names
arthro_trans <- arthro_trans [-1,] #remove first row (now redundant)
arthro_trans[, -1] <- apply(arthro_trans[, -1 , drop = FALSE], 2, as.numeric) #make data frame numeric
arthro_trans <- arthro_trans[,which(specnumber(arthro_trans,MARGIN=2)>0)] #remove singletons
arthro_binary <- ifelse(arthro_trans[, -1] > 0,1,0) #turn to binary data


metadata$site <- as.factor(metadata$site) #these are needed for plotting issues below 
metadata$SEASON <- as.factor(metadata$SEASON)
metadata$day <- as.factor(metadata$day)
#metadata$LAT <- as.numeric(metadata$LAT)
#metadata$LONG <- as.numeric(metadata$LONG)

arthroB_dist <- vegdist(arthro_binary, method = "bray") #distance marix of binary data using Bray_curtis method
arthro.nmds <- metaMDS(arthroB_dist) #run meta nmds analysis
permanova_season <- adonis2(arthroB_dist ~ SEASON, data = metadata)
permanova_season
season_tree <- nj(arthroB_dist)
ggtree(season_tree, layout = "rectangular") %<+% metadata +
  geom_tiplab(aes(colour = SEASON)) + 
  theme(legend.position = "right")

```

```{r}

plot(arthro.nmds, display = 'sites', type = 'n') ; points(arthro.nmds, display = 'sites', pch = 20) ; with(metadata, ordiellipse(arthro.nmds, metadata$SEASON, draw = 'polygon', col = 'blue', label = FALSE, show.groups =(c("WET")))) ; with(metadata, ordiellipse(arthro.nmds, metadata$SEASON, draw = 'polygon', col = 'goldenrod', label = FALSE,show.groups =(c("DRY"))))

season_tree <- nj(arthroB_dist)
ggtree(season_tree, layout = "rectangular") %<+% metadata +
  geom_tiplab(aes(colour = SEASON)) + 
  theme(legend.position = "right")

```

```{r}
arthro_wet <- subset(arthro_binary, metadata$SEASON == "WET")
metadata_wet <- subset(metadata, metadata$SEASON == "WET")
arthro_wet_dist <- vegdist(arthro_wet, method = 'bray')
arthro_wet.nmds <- metaMDS(arthro_wet_dist)
permanova_wet <- adonis2(arthro_wet_dist ~ day+site, data = metadata_wet)
summary(permanova_wet)
permanova_wet

permanova_wet_site <- adonis2(arthro_wet_dist ~ site, data = metadata_wet)
permanova_wet_site
```

```{r}
plot(arthro_wet.nmds, display = 'sites', type = 'n'); points(arthro_wet.nmds, display = 'sites', pch = 20,col = ifelse(metadata_wet$day == "A", "black", "red" ));with(metadata_wet, ordihull(arthro_wet.nmds, day, draw = 'polygon', col = c('black'), label = FALSE, show.groups =(c("A"))));with(metadata_wet, ordihull(arthro_wet.nmds, day, draw = 'polygon', col = c('red'), label = FALSE,show.groups =(c("B")))); with(metadata_wet, text(arthro_wet.nmds, labels = metadata_wet$site,pos = 2, cex = 0.5))

wet_tree <- nj(arthro_wet_dist)
ggtree(wet_tree, layout = "rectangular") %<+% metadata +
  geom_tiplab(aes(colour = day)) + 
  theme(legend.position = "right")


```

```{r}
arthro_dry <- subset(arthro_binary, metadata$SEASON == "DRY")
metadata_dry <- subset(metadata, metadata$SEASON == "DRY")
arthro_dry_dist <- vegdist(arthro_dry, method = 'bray')
arthro_dry.nmds <- metaMDS(arthro_dry_dist)
permanova_dry<- adonis2(arthro_dry_dist ~ day, data = metadata_dry)
summary(permanova_dry)
permanova_dry

permanova_dry_sites<- adonis2(arthro_dry ~ site, data = metadata_dry)
summary(permanova_dry_sites)
permanova_dry_sites
```

```{r}
plot(arthro_dry.nmds, display = 'sites', type = 'n'); points(arthro_dry.nmds, display = 'sites', pch = 20,col = ifelse(metadata_dry$day == "A", "black", "red" )); with(metadata_dry, ordihull(arthro_dry.nmds, day, draw = 'polygon', col = c('black'), label = FALSE, show.groups =(c("A")))); with(metadata_dry, ordihull(arthro_dry.nmds, day, draw = 'polygon', col = c('red'), label = FALSE, show.groups =(c("B")))); with(metadata_dry, text(arthro_dry.nmds, labels = metadata_dry$site,pos = 2, cex = 0.5))

dry_tree <- nj(arthro_dry_dist)
ggtree(dry_tree, layout = "rectangular") %<+% metadata +
  geom_tiplab(aes(colour = day)) + 
  theme(legend.position = "right")
```


```{r}

focal_groups <- c("f__Erebidae", "f__Crambidae", "f__Curculionidae", "f__Flatidae", "f__Formicidae", "f__Geometridae",
                  "f__Halictidae", "o__Blattodea", "f__Passalidae", "f__Pyralidae", "f__ Reduviidae", "f__Saturnidae",
                  "f__Scarabaeidae","f__Papilionidae", "f__Nymphalidae", "f__Lycaenidae", "f__Hesperiidae", "f__Apidae")

pattern <- paste0(focal_groups, collapse = "|")

arthro_all <- read.csv('data/correct_anal/tradmetabrparallel_final.csv')
metadata_all <- read.csv('data/correct_anal/metadata_trad_metabr_parallel.csv ')
metadata_samples <- filter(metadata_all, metadata_all$method == "LT") #remove from metadata file anythying not from light traps
metadata_samples <- metadata_samples %>% filter (location_good != 'other') %>% filter(!grepl("REP", sampleID)) #remove samples not from forestGEO plot

#we will already filter the data to include only LT from forestGEO sites 
metadata_samples$sampleID <- as.character(metadata_samples$sampleID) #sampleID columns as character

focal_arthro <- arthro_all[grep(pattern, arthro_all$classification), ]

colnames(focal_arthro) <- as.character(colnames(focal_arthro)) #column names (samples) as character
slected_cols <- intersect(colnames(focal_arthro), metadata_samples$sampleID) #intersect matches the names in sampleID to our columns
include_cols <- c(colnames(focal_arthro)[1:2], slected_cols)
focal_arthro_samples <- focal_arthro[, include_cols] #remember to include first two columns too since they are not in sampleID

focal_arthro_trans <- t(focal_arthro_samples[, -2]) #transpose dataset (samples in rows, species in columns)
colnames(focal_arthro_trans) <- focal_arthro_trans[1,] #first row is names
focal_arthro_trans <- focal_arthro_trans [-1,] #remove first row (now redundant)
focal_arthro_trans[, -1] <- apply(focal_arthro_trans[, -1 , drop = FALSE], 2, as.numeric) #make data frame numeric
focal_arthro_trans <- focal_arthro_trans[,which(specnumber(focal_arthro_trans,MARGIN=2)>0)] #remove singletons
focal_arthro_binary <- ifelse(focal_arthro_trans[, -1] > 0,1,0) #no need since tradmetabrparallel_final.csv already binary
#write.csv(focal_arthro_binary, 'data/correct_anal/focal_arthro_bianry.csv')

arthroall_dist <- vegdist(focal_arthro_binary, method = "bray") #distance marix of binary data using Jaccard method

#write.csv(arthro_all_binary, "./arthro_all_binary.csv")

metadata_samples$methodClass <- as.factor(metadata_samples$methodClass) #these are needed for plotting issues below 
metadata_samples$season <- as.factor(metadata_samples$season)
metadata_samples$location_good <- as.factor(metadata_samples$location_good)

arthroall_dist <- vegdist(focal_arthro_binary, method = "bray") #distance marix of binary data using Jaccard method
arthro_all.nmds <- metaMDS(arthroall_dist) #run meta nmds analysis
permanova_all <- adonis2(arthroall_dist ~ methodClass, data = metadata_samples)
summary(permanova_all)
permanova_all

permanova_all_sites <- adonis2(arthroall_dist ~ location_good, data = metadata_samples)
summary(permanova_all_sites)
permanova_all_sites

plot(arthro_all.nmds, display = 'sites', type = 'n')
points(arthro_all.nmds, display = 'sites',  pch = 20,col = ifelse(metadata_samples$season == "wet", "darkblue", "brown4" ))
with(metadata_samples, ordihull(arthro_all.nmds, methodClass,draw =  'polygon',col = c('firebrick4'),
                                  show.groups =(c("trad"))))
with(metadata_samples, ordispider(arthro_all.nmds, methodClass,draw =  'polygon',col = c('firebrick4'),
                                  show.groups =(c("trad"))))
with(metadata_samples, ordihull(arthro_all.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), 
                                  show.groups =(c("metabr"))))
with(metadata_samples, ordispider(arthro_all.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), 
                                  show.groups =(c("metabr"))))
with(metadata_samples, ordihull(arthro_all.nmds, methodClass, draw =  'polygon', col = c('darkorchid'), 
                                  show.groups =(c("parallel"))))
with(metadata_samples, ordispider(arthro_all.nmds, methodClass, draw =  'polygon', col = c('darkorchid'), 
                                  show.groups =(c("parallel"))))

all_tree <- nj(arthroall_dist)
ggtree(all_tree, layout = "rectangular") %<+% metadata_all +
  geom_tiplab(aes(colour = methodClass)) + 
  theme(legend.position = "right")


```

```{r}
arthro_all_wet <- read.csv('data/correct_anal/arthro_all_wet.csv', row.names = 1)

metadata_samples_wet <- subset(metadata_samples, metadata_samples$season == "wet")
arthro_all_wet_dist <- vegdist(arthro_all_wet, method = 'bray')
arthro_all_wet.nmds <- metaMDS(arthro_all_wet_dist)
permutation_all_wet <- adonis2(arthro_all_wet_dist ~ methodClass, data = metadata_samples_wet)
permutation_all_wet

permutation_all_wet_sites <-adonis2(arthro_all_wet_dist ~location_good, data = metadata_samples_wet)
permutation_all_wet_sites
```

```{r}
plot(arthro_all_wet.nmds, display = 'sites', type = 'n')
points(arthro_all_wet.nmds, display = 'sites', pch = 20)
with(metadata_samples_wet, ordispider(arthro_all_wet.nmds, methodClass, draw =  'polygon', col = c('firebrick4'), label = F,show.groups =(c("trad"))))
with(metadata_samples_wet, ordihull(arthro_all_wet.nmds, methodClass, draw =  'polygon', col = c('firebrick4'), label = F,show.groups =(c("trad"))))
with(metadata_samples_wet, ordispider(arthro_all_wet.nmds, methodClass,draw =  'polygon', col = c('darkolivegreen'), label = FALSE,show.groups =(c("metabr"))))
with(metadata_samples_wet, ordihull(arthro_all_wet.nmds, methodClass,draw =  'polygon', col = c('darkolivegreen'), label = FALSE,show.groups =(c("metabr"))))
with(metadata_samples_wet, ordispider(arthro_all_wet.nmds, methodClass,draw =  'polygon', col = c('darkorchid'), label = FALSE,show.groups =(c("parallel"))))
with(metadata_samples_wet, ordihull(arthro_all_wet.nmds, methodClass,draw =  'polygon', col = c('darkorchid'), label = FALSE,show.groups =(c("parallel"))))
with(metadata_samples_wet, text(arthro_all_wet.nmds, labels = metadata_samples_wet$location_good,pos = 2, cex = 0.4))

all_wet_tree <- nj(arthro_all_wet_dist)
ggtree(all_wet_tree, layout = "rectangular") %<+% metadata_samples_wet +
  geom_tiplab(aes(colour = methodClass)) + 
  theme(legend.position = "right")

metadata_samples_wet$combined_group <- ifelse(metadata_samples_wet$methodClass %in% c("trad", "parallel"), "trad_parallel", metadata_samples_wet$methodClass)

# Plot NMDS with combined polygon for "trad" and "parallel"
plot(arthro_all_wet.nmds, display = 'sites', type = 'n')
points(arthro_all_wet.nmds, display = 'sites', pch = 20)
with(metadata_samples_wet, ordispider(arthro_all_wet.nmds, combined_group, draw = 'polygon', col = c('firebrick4'), label = FALSE, show.groups = c("trad_parallel")))
with(metadata_samples_wet, ordihull(arthro_all_wet.nmds, combined_group, draw = 'polygon', col = c('firebrick4'), label = FALSE, show.groups = c("trad_parallel")))
with(metadata_samples_wet, ordispider(arthro_all_wet.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), label = FALSE, show.groups =(c("metabr"))))
with(metadata_samples_wet, ordihull(arthro_all_wet.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), label = FALSE, show.groups =(c("metabr"))))
with(metadata_samples_wet, text(arthro_all_wet.nmds, labels = metadata_samples_wet$location_good, pos = 2, cex = 0.4))



```

```{r}
arthro_all_dry <- read.csv('data/correct_anal/arthro_all_dry.csv', row.names = 1)
metadata_samples_dry <- subset(metadata_samples, metadata_samples$season == "dry")
arthro_all_dry_dist <- vegdist(arthro_all_dry, method = 'bray')
arthro_all_dry.nmds <- metaMDS(arthro_all_dry_dist)

permutation_all_dry <- adonis2(arthro_all_dry_dist ~ methodClass+location_good, data = metadata_samples_dry)
permutation_all_dry

permutation_all_dry_sites <-adonis2(arthro_all_dry_dist ~location_good, data = metadata_samples_dry)
permutation_all_dry_sites
```

```{r}
plot(arthro_all_dry.nmds, display = 'sites', type = 'n')
points(arthro_all_dry.nmds, display = 'sites', pch = 20)
with(metadata_samples_dry, ordispider(arthro_all_dry.nmds, methodClass, draw =  'polygon',  col = c('firebrick4'), label = FALSE, show.groups =(c("trad"))))
with(metadata_samples_dry, ordihull(arthro_all_dry.nmds, methodClass, draw =  'polygon',  col = c('firebrick4'), label = FALSE, show.groups =(c("trad"))))
with(metadata_samples_dry, ordispider(arthro_all_dry.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), label = FALSE, show.groups =(c("metabr"))))
with(metadata_samples_dry, ordihull(arthro_all_dry.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), label = FALSE, show.groups =(c("metabr"))))
with(metadata_samples_dry, ordispider(arthro_all_dry.nmds, methodClass, draw =  'polygon',  col = c('darkorchid'), label = FALSE, show.groups =(c("parallel"))))
with(metadata_samples_dry, ordihull(arthro_all_dry.nmds, methodClass, draw =  'polygon',  col = c('darkorchid'), label = FALSE, show.groups =(c("parallel"))))
with(metadata_samples_dry, text(arthro_all_dry.nmds, labels = metadata_samples_dry$location_good,pos = 2, cex = 0.4))

all_dry_tree <- nj(arthro_all_dry_dist)
ggtree(all_dry_tree, layout = "rectangular") %<+% metadata_samples_dry +
  geom_tiplab(aes(colour = methodClass)) + 
  theme(legend.position = "right")






metadata_samples_dry$combined_group <- ifelse(metadata_samples_dry$methodClass %in% c("trad", "parallel"), "trad_parallel", metadata_samples_dry$methodClass)

# Plot NMDS with combined polygon for "trad" and "parallel"
plot(arthro_all_dry.nmds, display = 'sites', type = 'n')
points(arthro_all_dry.nmds, display = 'sites', pch = 20)
with(metadata_samples_dry, ordispider(arthro_all_dry.nmds, combined_group, draw = 'polygon', col = c('firebrick4'), label = FALSE, show.groups = c("trad_parallel")))
with(metadata_samples_dry, ordihull(arthro_all_dry.nmds, combined_group, draw = 'polygon', col = c('firebrick4'), label = FALSE, show.groups = c("trad_parallel")))
with(metadata_samples_dry, ordispider(arthro_all_dry.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), label = FALSE, show.groups =(c("metabr"))))
with(metadata_samples_dry, ordihull(arthro_all_dry.nmds, methodClass, draw =  'polygon', col = c('darkolivegreen'), label = FALSE, show.groups =(c("metabr"))))
with(metadata_samples_dry, text(arthro_all_dry.nmds, labels = metadata_samples_dry$location_good, pos = 2, cex = 0.4))


```
